# docs/openapi (1).json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BAS Orchestrator API",
    "description": "HTTP trigger surface for the BAS internal-attack orchestrator.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/engagements": {
      "post": {
        "summary": "Submit Engagement",
        "operationId": "submit_engagement_engagements_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngagementCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Engagements",
        "operationId": "list_engagements_engagements_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EngagementSummary"
                  },
                  "title": "Response List Engagements Engagements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/results": {
      "post": {
        "tags": [
          "results"
        ],
        "summary": "Receive execution results from BAS backend",
        "description": "Accept execution results POSTed by the BAS backend after abilities are executed on the target agent. The `engagement_id` (from `POST /engagements`) and `operation_id` are both in the request body.\n\n**Idempotent**: duplicate `operation_id` returns 200 without re-processing.",
        "operationId": "receive_results_results_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationResultRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultAcceptedResponse"
                }
              }
            }
          },
          "200": {
            "description": "Duplicate result (already received)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultAcceptedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Engagement not found"
          },
          "409": {
            "description": "Engagement not in awaiting_results or running status"
          },
          "422": {
            "description": "Invalid JSON or missing/invalid fields"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AbilityResultRequest": {
        "properties": {
          "ability_id": {
            "type": "string",
            "title": "Ability Id",
            "description": "The ability UUID that was pushed to the BAS backend."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Ability name.",
            "default": ""
          },
          "mitre_technique_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mitre Technique Id",
            "description": "MITRE ATT&CK technique ID, e.g. T1016."
          },
          "platform": {
            "type": "string",
            "title": "Platform",
            "description": "Target platform: windows, linux, mac.",
            "default": ""
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageExecutionRequest"
            },
            "type": "array",
            "title": "Stages",
            "description": "List of stage execution results."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "ability_id"
        ],
        "title": "AbilityResultRequest",
        "description": "Execution outcome for a single ability (all its stages)."
      },
      "EngagementCreateRequest": {
        "properties": {
          "phases": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Phase"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phases",
            "description": "Kill-chain phases to execute, in order. Accepted values: discovery, privesc, credaccess, lateral, persistence, defevasion, impact. Aliases like 'recon', 'priv-esc', 'credential-access' are also accepted and normalised automatically. When omitted, the LLM router decides each step.",
            "examples": [
              [
                "discovery"
              ],
              [
                "discovery",
                "privesc"
              ]
            ]
          },
          "environment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EnvironmentSelector"
              },
              {
                "type": "null"
              }
            ],
            "description": "Which BAS environment to use. Defaults to most-recent."
          },
          "target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TargetHint"
              },
              {
                "type": "null"
              }
            ],
            "description": "Preferences for picking the foothold agent."
          },
          "max_iterations": {
            "type": "integer",
            "maximum": 50,
            "minimum": 1,
            "title": "Max Iterations",
            "description": "Hard cap on master-router iterations across the engagement. Each iteration is one master phase decision (briefing or commit); internal planner attempts (≤ 3 per phase) and master revisions (≤ 1 per phase) are capped separately.",
            "default": 20
          },
          "dry_run": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dry Run",
            "description": "Override `bas.dry_run` from config when set."
          }
        },
        "type": "object",
        "title": "EngagementCreateRequest",
        "description": "Minimum trigger payload — everything else is auto-resolved."
      },
...
      "EngagementSubmitResponse": {
        "properties": {
          "engagement_id": {
            "type": "string",
            "title": "Engagement Id"
          },
          "status_url": {
            "type": "string",
            "title": "Status Url"
          },
          "status": {
            "type": "string",
            "const": "queued",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "engagement_id",
          "status_url",
          "status"
        ],
        "title": "EngagementSubmitResponse"
      },
      "EngagementSummary": {
        "properties": {
          "engagement_id": {
            "type": "string",
            "title": "Engagement Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "awaiting_results",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "awaiting_since": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awaiting Since"
          },
          "iterations": {
            "type": "integer",
            "title": "Iterations"
          },
          "completed_stages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Completed Stages"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "engagement_id",
          "status",
          "started_at",
          "finished_at",
          "iterations",
          "completed_stages"
        ],
        "title": "EngagementSummary"
      },
      "EnvironmentSelector": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "BAS environment_id (UUID)."
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Exact environment name."
          }
        },
        "type": "object",
        "title": "EnvironmentSelector",
        "description": "Optional pin for which BAS environment to use. Defaults to most-recent."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OperationResultRequest": {
        "properties": {
          "engagement_id": {
            "type": "string",
            "title": "Engagement Id",
            "description": "The engagement this result belongs to (returned by POST /engagements).",
            "examples": [
              "08c047ac0cba49b3ae85257406a5cc28"
            ]
          },
          "operation_id": {
            "type": "string",
            "title": "Operation Id",
            "description": "Unique UUID for this execution run.",
            "examples": [
              "f9e8d7c6-b5a4-3210-fedc-ba0987654321"
            ]
          },
          "operation_name": {
            "type": "string",
            "title": "Operation Name",
            "description": "Human-readable operation label.",
            "default": ""
          },
          "operation_status": {
            "type": "string",
            "enum": [
              "completed",
              "failed",
              "partial"
            ],
            "title": "Operation Status",
            "description": "Overall operation outcome.",
            "default": "completed"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At",
            "description": "When execution finished (ISO 8601)."
          },
          "adversary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Adversary",
            "description": "Adversary metadata: {adversary_id, name}."
          },
          "progress": {
            "additionalProperties": true,
            "type": "object",
            "title": "Progress",
            "description": "Execution progress: {total_abilities, completed_abilities, progress_percent}."
          },
          "abilities": {
            "items": {
              "$ref": "#/components/schemas/AbilityResultRequest"
            },
            "type": "array",
            "title": "Abilities",
            "description": "List of ability execution results."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "engagement_id",
          "operation_id"
        ],
        "title": "OperationResultRequest",
        "description": "Top-level result payload POSTed by the BAS backend after execution."
      },
      "Phase": {
        "type": "string",
        "enum": [
          "discovery",
          "privesc",
          "credaccess",
          "lateral",
          "persistence",
          "defevasion",
          "impact"
        ],
        "title": "Phase",
        "description": "Canonical kill-chain phases — the contract between backend and orchestrator.\n\nBeing a ``str`` enum, each member compares equal to its plain-string value:\n``Phase.DISCOVERY == \"discovery\"`` is ``True``."
      },
...
      "ResultAcceptedResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "already_received"
            ],
            "title": "Status"
          },
          "operation_id": {
            "type": "string",
            "title": "Operation Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "operation_id"
        ],
        "title": "ResultAcceptedResponse"
      },
...
      "StageExecutionRequest": {
        "properties": {
          "stage_name": {
            "type": "string",
            "title": "Stage Name",
            "description": "Stage identifier."
          },
          "executor": {
            "type": "string",
            "title": "Executor",
            "description": "Executor used: cmd, psh, sh, bash.",
            "default": ""
          },
          "command_executed": {
            "type": "string",
            "title": "Command Executed",
            "description": "The actual command that was run.",
            "default": ""
          },
          "execution_status": {
            "type": "string",
            "enum": [
              "passed",
              "failed"
            ],
            "title": "Execution Status",
            "description": "Whether the stage passed or failed.",
            "default": "failed"
          },
          "stdout": {
            "type": "string",
            "title": "Stdout",
            "description": "Full standard output captured.",
            "default": ""
          },
          "stderr": {
            "type": "string",
            "title": "Stderr",
            "description": "Full standard error captured.",
            "default": ""
          },
          "exit_code": {
            "type": "integer",
            "title": "Exit Code",
            "description": "Exit code. 0 = success.",
            "default": -1
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp",
            "description": "When this stage executed (ISO 8601)."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "stage_name"
        ],
        "title": "StageExecutionRequest",
        "description": "One executed stage within an ability."
      },
      "TargetHint": {
        "properties": {
          "platform": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "windows",
                  "linux",
                  "mac"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform"
          }
        },
        "type": "object",
        "title": "TargetHint",
        "description": "Optional preferences when selecting the foothold agent."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}